Feature1: Expression observable.
This component represents an observable that allows for subscription to streams of expressions (strings).
The source of this stream is a BufferedReader that wraps the true source of data (e.g. a stream created
from the output of a socket, channel, etc.).

First version of the solution class diagram

There are still some important classes missing (like a mediator that mediates between a resource loader
and the drawing tool, and the AbstractDrawing command that allows for specification of the symbol used
to draw.

Replaced rx-async with rx-string to use InputStream/Reader with Observables

Added guava & rx-file dependencies and a utility class for file i/o

Utility class contains methods to open a reader for text files using the strategy chain:
classpath, file system, and url.

TODO make those classes Strategy classes, and the chain a Chain of Responsibility.

Added test for StringObservable#byLine

This observable is one of the two alternatives for implementing the ExpressionObserver.
It uses blocking i/o to read from the file.The other option is the FileObservable,
which uses non-blocking i/o).

Added test for StringObservable#map

Test maps string expressions to mock classes representing commands. The next step
would be to add the Canvas as a subscriber for those commands.

Minor refactor to StringObservableTest class to promote code legibility

Adding Observer that streams expressions from a file. Tests are still pending.

This is an emergency commit; just to enable working from home.

Added tests for ExpressionObservable class

A major refactor in testing code had to be performed to be able to implement the tests for
this class in a legible and maintainable way.

Added a test to validate what happens when observables use invalid readers

The test checks that subscribers are notified of the error.

Added a test that validates what happens when observers unsubscribe.

A new class was created to simulate unsubscription after a configurable amount of received
items.