Feature 2: ObjectFactory & Feature 3: Drawing Expressions
Added basic factory abstract factory implementation and a test for happy path

Currently the abstract factory only returns one of the components that make part of the solution.
Pending tests have to be implemented on the factory to guarantee it behaves correctly.

Completed tests for ObjectFactory implementation

ObjectFactory implementation had to be modified as a result of some tests that proved its
initial implementation was faulty.

Added new method to ObjectFactory that allows for expression creation

Added tests for CanvasExpression creation

Added tests for LineExpression creation

Added tests for RectangleExpression creation

Added tests for BucketFillExpression creation