Feature 5: Interpret CanvasExpression to paint Canvas
Added method to Canvas object that allows it to paint itself to an OutputStream

Added a method to de DrawingToolsUtility class that allows setup of an output file configured in
the application configuration properties; the method will return System.out if no output file is
configured.

Feature 5: Interpret CanvasExpression to paint Canvas.